refactor: 💡 Extract the basic arguments, and rename file
Refactor to be able to import the basic arguments in multiple
(independed) scripts
