+++ src/cmd/apply.ts
@@ -3,10 +3,10 @@ import { Argv, CommandModule } from 'yargs'
 import { $ } from 'zx'
 import { OtomiDebugger, terminal } from '../common/debug'
 import { giteaPush } from '../common/gitea-push'
-import { Arguments as HelmArgs, helmOptions } from '../common/helm-opts'
 import { hf, hfStream } from '../common/hf'
 import { ENV, LOG_LEVEL_STRING } from '../common/no-deps'
 import { cleanupHandler, otomi, PrepareEnvironmentOptions } from '../common/setup'
+import { Arguments as HelmArgs, helmOptions } from '../common/yargs-opts'
 import { ProcessOutputTrimmed } from '../common/zx-enhance'
 import { decrypt } from './decrypt'
 import { Arguments as DroneArgs, genDrone } from './gen-drone'
