+++ src/cmd/pull.ts
@@ -1,9 +1,9 @@
 import { Argv } from 'yargs'
 import { $ } from 'zx'
 import { OtomiDebugger, terminal } from '../common/debug'
-import { Arguments as HelmArgs, helmOptions } from '../common/helm-opts'
 import { ENV } from '../common/no-deps'
 import { cleanupHandler, otomi, PrepareEnvironmentOptions } from '../common/setup'
+import { Arguments as HelmArgs, helmOptions } from '../common/yargs-opts'
 import { Arguments as BootsrapArgs, bootstrap } from './bootstrap'
 
 interface Arguments extends HelmArgs, BootsrapArgs {}
