+++ src/cmd/score-templates.ts
@@ -2,10 +2,10 @@ import { unlinkSync, writeFileSync } from 'fs'
 import { Argv } from 'yargs'
 import { $, nothrow } from 'zx'
 import { OtomiDebugger, terminal } from '../common/debug'
-import { Arguments, helmOptions } from '../common/helm-opts'
 import { hfTemplate } from '../common/hf'
 import { ENV } from '../common/no-deps'
 import { cleanupHandler, otomi, PrepareEnvironmentOptions } from '../common/setup'
+import { Arguments, helmOptions } from '../common/yargs-opts'
 
 const fileName = 'score-template'
 const templatePath = '/tmp/template.yaml'
