+++ src/cmd/sync.ts
@@ -1,9 +1,9 @@
 import { Argv } from 'yargs'
 import { OtomiDebugger, terminal } from '../common/debug'
-import { Arguments, helmOptions } from '../common/helm-opts'
 import { hfStream } from '../common/hf'
 import { ENV, LOG_LEVEL_STRING } from '../common/no-deps'
 import { cleanupHandler, otomi, PrepareEnvironmentOptions } from '../common/setup'
+import { Arguments, helmOptions } from '../common/yargs-opts'
 import { decrypt } from './decrypt'
 
 const fileName = 'sync'
